+++ src/main/scala/io/github/dexclaimation/subpub/SubPub.scala
@@ -9,7 +9,7 @@ package io.github.dexclaimation.subpub
 
 import akka.actor.typed.scaladsl.AskPattern._
 import akka.actor.typed.{ActorRef, ActorSystem, Props, SpawnProtocol}
-import akka.stream.scaladsl.Source
+import akka.stream.scaladsl.{Concat, Source}
 import akka.util.Timeout
 import akka.{Done, NotUsed}
 import io.github.dexclaimation.subpub.model.SubIntent
@@ -18,8 +18,15 @@ import scala.concurrent.duration._
 import scala.concurrent.{Await, ExecutionContext, Future}
 import scala.reflect.ClassTag
 
+/**
+ * Pub/Sub system for topic based distributed streams
+ *
+ * @param bufferSize    Buffer size for the upstream producer.
+ * @param idleDuration  The amount of time used to wait for the SubPub to set up.
+ * @param awaitDuration The amount of time to wait for Source.
+ */
 class SubPub(
-  bufferSize: Int = 100,
+  bufferSize: Int = 256,
   idleDuration: FiniteDuration = 3.seconds,
   awaitDuration: FiniteDuration = 5.seconds,
 )(implicit system: ActorSystem[SpawnProtocol.Command]) {
@@ -72,6 +79,71 @@ class SubPub(
     .result(asyncSource[T](topic), awaitDuration)
 
   /**
+   * Synchronously request for an existing topic stream or create a new one.
+   *
+   * @param topic     The topic used to differentiate streams.
+   * @param initValue The initial value pushed into the stream.
+   * @tparam T The type for the Source
+   * @return Source of type T.
+   */
+  def source[T: ClassTag](topic: String, initValue: T): Source[T, NotUsed] =
+    combineSource(source[T](topic),
+      init = Some(Source.single(initValue)),
+      end = None
+    )
+
+  /**
+   * Synchronously request for an existing topic stream or create a new one.
+   *
+   * @param topic    The topic used to differentiate streams.
+   * @param endValue The ending value pushed into the stream.
+   * @tparam T The type for the Source
+   * @return Source of type T.
+   */
+  def source[T: ClassTag](topic: String, endValue: => T): Source[T, NotUsed] =
+    combineSource(source[T](topic),
+      init = None,
+      end = Some(Source.single(endValue))
+    )
+
+  /**
+   * Synchronously request for an existing topic stream or create a new one.
+   *
+   * @param topic     The topic used to differentiate streams.
+   * @param initValue The initial value pushed into the stream.
+   * @param endValue  The ending value pushed into the stream.
+   * @tparam T The type for the Source
+   * @return Source of type T.
+   */
+  def source[T: ClassTag](topic: String, initValue: T, endValue: T): Source[T, NotUsed] =
+    combineSource(source[T](topic),
+      init = Some(Source.single(initValue)),
+      end = Some(Source.single(endValue))
+    )
+
+  /**
+   * Combine source with an initial and/or an ending one
+   *
+   * @param stream Core downstream.
+   * @param init   Initial downstream values.
+   * @param end    Ending downstream values.
+   * @return Combined source that push the initial (if exist) first, then core, then ending (if exist).
+   */
+  private def combineSource[T: ClassTag](
+    stream: Source[T, NotUsed],
+    init: Option[Source[T, NotUsed]],
+    end: Option[Source[T, NotUsed]]
+  ): Source[T, NotUsed] = (init, end) match {
+    case (Some(stream1), Some(stream2)) =>
+      Source.combine(stream1, stream, stream2)(Concat(_))
+    case (Some(stream1), None) =>
+      Source.combine(stream1, stream)(Concat(_))
+    case (None, Some(stream2)) =>
+      Source.combine(stream, stream2)(Concat(_))
+    case _ => stream
+  }
+
+  /**
    * Subscribe for a specific topic.
    *
    * @param topic      The topic for what to subscribe to.
@@ -96,9 +168,24 @@ class SubPub(
   /**
    * Close a topic source.
    *
-   * @param topic    The topic to be ended
-   * @param forceFul Kill forcefully or complete it.
+   * @param topic The topic to be ended
    */
   def close(topic: String): Unit =
     engine ! SubIntent.AcidPill(topic)
+}
+
+object SubPub {
+
+  /**
+   * Pub/Sub system for topic based distributed streams
+   *
+   * @param bufferSize    Buffer size for the upstream producer.
+   * @param idleDuration  The amount of time used to wait for the SubPub to set up.
+   * @param awaitDuration The amount of time to wait for Source.
+   */
+  def apply(
+    bufferSize: Int = 256,
+    idleDuration: FiniteDuration = 3.seconds,
+    awaitDuration: FiniteDuration = 5.seconds,
+  )(implicit system: ActorSystem[SpawnProtocol.Command]): SubPub = new SubPub(bufferSize, idleDuration, awaitDuration)
 }