+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+### v0.1.7
+
+- Added `source` method with `initValue` and / or `endValue` to have a `Source` with a initial data or closing one (Only
+  for that source not the topic).
+- Added `apply` for `SubPub` to construct without the `new` keyword.
+- Removed unused code.
+- Updated default buffer size to `256`.
+- Updated package description to the actual github repo.
+
 ### v0.1.6
 
 - Fixed issue of faulty math logic for `BroadcastHub` buffer size.